+++ lib/td_cx/jobs/search.ex
@@ -90,6 +90,6 @@ defmodule TdCx.Jobs.Search do
   end
 
   defp simple_query_string_clause(%{simple_search_fields: fields}) do
-    %{simple_query_string: %{fields: fields}}
+    %{simple_query_string: %{fields: fields, quote_field_suffix: ".exact"}}
   end
 end