+++ test/td_dd_web/controllers/grant_search_controller_test.exs
@@ -62,10 +62,11 @@ defmodule TdDdWeb.GrantSearchControllerTest do
                      simple_query_string: %{
                        fields: [
                          "user.full_name",
-                         "data_structure_version.name*",
-                         "data_structure_version.original_name*"
+                         "data_structure_version.name",
+                         "data_structure_version.original_name"
                        ],
-                       query: "\"foo\""
+                       query: "\"foo\"",
+                       quote_field_suffix: ".exact"
                      }
                    },
                    must_not: %{exists: %{field: "deleted_at"}}